# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "User registered succesfully, Use Login method to receive token.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Please see errors parameter for all errors.",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Detail error description",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login User",
                "description": "User Login here",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "User registered succesfully, Use Login method to receive token.",
                                            "type": "string"
                                        },
                                        "token": {
                                            "description": "Use this token for other requests.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Error general description.",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Detail error description",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User authentication failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Error general description.",
                                            "type": "string"
                                        },
                                        "errors": {
                                            "description": "Detail error description",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-detail": {
            "post": {
                "tags": [
                    "User detail"
                ],
                "summary": "user-detail",
                "description": "Get user details",
                "operationId": "user-detail",
                "responses": {
                    "200": {
                        "description": "User detail fetched in Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "false",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "Data fetched successfully",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "User list.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "logout",
                "description": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "User logout successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "true|false",
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "User logout successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}